feat(refactor-frontend-errorHandling): Implement error handling at the frontend
- [x] refactor error handling
- [x] modify seeder

[Finishes #169297433]
