+++ CLIENT/actions/bank/addUserBankDetails.js
@@ -24,20 +24,19 @@ const addUserBankDetails = () => {
         bvn,
       }),
     }).then(res => res.json()).then((data) => {
-      console.log(data);
       if (data.status === 409 && data.error) {
-        document.getElementById('msg').innerHTML = data.error;
-        open('backdrop2', 'errorBox');
+        throw new Error(data.error);
       } else if (data.status === 422 && data.errors) {
-        document.getElementById('msg').innerHTML = data.errors[0].error;
-        open('backdrop2', 'errorBox');
+        throw new Error(data.errors[0].error);
       } else if (data.status === 404) {
-        document.getElementById('msg').innerHTML = data.error;
-        open('backdrop2', 'errorBox');
+        throw new Error(data.error);
       } else {
         displayBankDetails(data.data);
         profileAlert('User bank details successfully added', '-green');
       }
+    }).catch((err) => {
+      document.getElementById('msg').innerHTML = err.message;
+      open('backdrop2', 'errorBox');
     });
   }
 };
