+++ CLIENT/actions/employment/addUserEmployment.js
@@ -31,23 +31,20 @@ const addUserEmployment = () => {
         companySector: sector,
       }),
     }).then(res => res.json()).then((data) => {
-      console.log(data);
       if (data.status === 409 && data.error) {
-        document.getElementById('msg').innerHTML = data.error;
-        open('backdrop2', 'errorBox');
+        throw new Error(data.error);
       } else if (data.status === 422 && data.errors) {
-        document.getElementById('msg').innerHTML = data.errors[0].error;
-        open('backdrop2', 'errorBox');
+        throw new Error(data.errors[0].error);
       } else if (data.status === 404) {
-        document.getElementById('msg').innerHTML = data.error;
-        open('backdrop2', 'errorBox');
+        throw new Error(data.error);
       } else {
         displayEmploymentData(data.data);
         profileAlert('User employment details successfully added', '-green');
       }
+    }).catch((err) => {
+      document.getElementById('msg').innerHTML = err.message;
+      open('backdrop2', 'errorBox');
     });
-  } else {
-    console.log(isUserEmploymentValid);
   }
 };
