+++ CLIENT/actions/loan/modifyLoanStatus/modifyLoanStatus.js
@@ -13,15 +13,15 @@ const modifyLoanStatus = (loanId, LoanStatus) => {
       const { id, status, client } = data.data;
       toggleLoanStatus({ id, status, client });
     } else if (data.status === 409) {
-      document.getElementById('msg').innerHTML = data.error;
-      open('backdrop1', 'errorBox');
+      throw new Error(data.error);
     } else if (data.status === 422) {
-      document.getElementById('msg').innerHTML = data.error;
-      open('backdrop1', 'errorBox');
+      throw new Error(data.error);
     } else if (data.status === 404) {
-      document.getElementById('msg').innerHTML = data.error;
-      open('backdrop1', 'errorBox');
+      throw new Error(data.error);
     }
+  }).catch((err) => {
+    document.getElementById('msg').innerHTML = err.message;
+    open('backdrop1', 'errorBox');
   });
 };
