+++ CLIENT/admin.html
@@ -101,7 +101,7 @@
                     </div>
                     <div class="dashboard__header-email-display -col-l-5 -col-m-8 -text-align-right -fc-gray hide-on-medium-and-down -fs-res-l-1x" id="username">Jane Doe</div>
                     <div class="dashboard__title -fs-res-l-1 -fc-white -android-lollipop-2">Admin Dashboard</div>
-                    <h4 class="dashboard__greet hide-on-large-only" id="greet">Welcome Jane</h4>
+                    <!-- <h4 class="dashboard__greet hide-on-large-only" id="greet">Welcome Jane</h4> -->
                     <div class="dashboard__body">
                             
                             <div class="card-container -card-align -card-container-top" id="cardContainer">
