+++ CLIENT/signup.html
@@ -60,6 +60,7 @@
                         <h1
                             class="sign__header -text-align-center">
                             Signup!</h1>
+                            <div id="errorMsg" class="-text-align-center"></div>
                         <div class="sign__body">
                             <center>
                                 <div class="-col-sm-11"><input type="text" id="firstname"
