+++ SERVER/src/seeder/data/data.js
@@ -3,8 +3,8 @@ const data = {
     firstname: 'admin',
     lastname: 'super',
     email: 'admin.super@gmail.com',
-    phone: '09078282176',
-    password: 'admin12345',
+    phone: '09078282177',
+    password: 'admin12346',
     isadmin: true,
   },
   users: [
